Add development font fallback
