+++ next-frontend/src/app/(wca)/layout.tsx
@@ -6,26 +6,14 @@ import { Provider as UiProvider } from "@/components/ui/provider";
 import Navbar from "./navbar";
 import Footer from "@/components/Footer";
 import RandomBackground from "@/components/RandomBackground";
-import { Rubik } from "next/font/google";
 import { ThemeProvider } from "@wrksz/themes/next";
+import { computeAppFont } from "@/styles/appFont";
 
 export const metadata: Metadata = {
   title: {
     template: "%s | World Cube Association",
     default: "",
   },
-};
-
-const devFont = Rubik({ subsets: ["latin"] });
-
-const computeFont = async () => {
-  if (process.env.PROPRIETARY_FONT === "TTNormsPro") {
-    const { ttNormsPro } = await import("@/styles/fonts");
-
-    return ttNormsPro;
-  }
-
-  return devFont;
 };
 
 export const dynamic = "force-dynamic";
@@ -35,7 +23,7 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const appFont = await computeFont();
+  const appFont = await computeAppFont();
 
   return (
     <html suppressHydrationWarning>
